Changed profiling UX #1043
